requirements.txt
robotframework==7.4.2
Appium-Python-Client==5.2.7
robotframework-appiumlibrary==3.2.1
